Updated branches of KDevelop and related packages
Commit 3435f36 updates kdev-valgrind to latest kdevplatform API,
but this breaks compatibility with older kdevplatform 5.3 API.
kdev-valgrind does provides a master branch but no version-
targeted branches for newer releases than 5.1, which is highly
outdated. kdesrc-build does not allow to choose for specific
commits, e.g. the 'revision' option is not respected. Thus the
only way to successfully compile kdev-valgrind is to update
kdevelop, including kdevplatform and plugins.