Replaced manual service setup with jm1.systemd role
Now service states are only changed if required. Previously, systemctl
was called even if unnecessary, which let Ansible always think that
the service states had been altered with every call.

The systemd role provided by Ansible is not used, because it requires
a system booted with systemd as init system for managing services,
e.g. it fails in Packer's docker builder: "System has not been booted
 with systemd as init system (PID 1). Can't operate. Failed to connect
 to bus: Host is down"